common makeSimpleFlat: take 'ind' as a 'unsigned'
Pays down some cast noise from prior commits.

+++ lib/dotgen/dotsplines.c
@@ -1166,7 +1166,7 @@ static void makeSimpleFlatLabels(node_t *tn, node_t *hn, edge_t **edges,
     free (earray);
 }
 
-static void makeSimpleFlat(node_t *tn, node_t *hn, edge_t **edges, int ind,
+static void makeSimpleFlat(node_t *tn, node_t *hn, edge_t **edges, unsigned ind,
                            unsigned cnt, int et) {
     edge_t* e = edges[ind];
     pointf points[10], tp, hp;
@@ -1179,7 +1179,7 @@ static void makeSimpleFlat(node_t *tn, node_t *hn, edge_t **edges, int ind,
     dy = tp.y - (cnt > 1 ? ND_ht(tn) / 2. : 0.);
 
     for (unsigned i = 0; i < cnt; i++) {
-	e = edges[ind + (int)i];
+	e = edges[ind + i];
 	size_t pointn = 0;
 	if (et == EDGETYPE_SPLINE || et == EDGETYPE_LINE) {
 	    points[pointn++] = tp;
@@ -1248,7 +1248,7 @@ static void make_flat_adj_edges(graph_t *g, edge_t **edges, unsigned ind,
     if (ports == 0) {
 	/* flat edges without ports and labels can go straight left to right */
 	if (labels == 0) {
-	    makeSimpleFlat(tn, hn, edges, (int)ind, cnt, et);
+	    makeSimpleFlat(tn, hn, edges, ind, cnt, et);
 	}
 	/* flat edges without ports but with labels take more work */
 	else {
@@ -1588,7 +1588,7 @@ static void make_flat_edge(graph_t *g, spline_info_t *sp, path *P,
     }
 
     if (et == EDGETYPE_LINE) {
-	makeSimpleFlat(agtail(e), aghead(e), edges, (int)ind, cnt, et);
+	makeSimpleFlat(agtail(e), aghead(e), edges, ind, cnt, et);
 	return;
     }
